fix and some developments

+++ MicroServiceProject/Source/Core/MP.Core.Domain/Entities/BaseEntity.cs
@@ -11,7 +11,7 @@ namespace MP.Core.Domain.Entities
     public class BaseEntity
     {
         [Key]
-        public long Id { get; set; } = 0; //Int primary key, 0,1,2,3,..
+        public int Id { get; set; } = 0; //Int primary key, 0,1,2,3,..
         public DateTime Created { get; set; } = DateTime.Now; //Created Time
         [Range(0, 2)]
         public int Status { get; set; } = (int)EntityStatus.None; //Status
